Swift: generalize open redirection on both platforms and on outputs
The hash map mechanism that was already in use for reading swiftmodule
files on macOS is now in use also on renames and on Linux. This replaces
the whole output flag rewrite mechanism.

Note: all writes and renames are also redirected in order to avoid our
extractor changing the output files of the original build.